Connect-K game in ARM and C
CONNECT-K

In this last practice, we are going to include all the logic to play Connect K. You must include the necessary code retrieved from the first practice. The functions to check the move must be those that provided the best performance in Practice 1.

It is essential to design the state machine and the input/output of the game to meet all the following requirements:

At the start of the game, the program will display information about the game and how to play via the serial line while waiting for any button press or a new command for a new game.
The acceptable commands encoded in ASCII are:
End the game: $END!
New game: $NEW!
Move: $#-#! where #-# will be the row-column number encoded with two characters (1..7, 1..7). Example: the command $2-3! indicates row 2 and column 3.
The input and output from the previous practices will be maintained unless modified in the following specifications.
The board will then appear on the serial line in the console; the format is left free, but you should try to make it look as good as possible to facilitate playing. In the first game, and to be able to debug and present the project, it will start with the test board from Practice 1 and the turn of player one.
It will indicate at each moment which player is to play.
We will include the arrival of commands via the serial line and maintain the options to accept or cancel via buttons developed in Practice 2. The row and column will be received via the serial line. If the received command or row-column is invalid, it will be ignored, and the message “Invalid column” or “Incorrect command” will be displayed on the screen, and the GPIO29 LED will light up (following the reservation and communication pattern to the module). Otherwise, the board will be displayed with a special character for the new piece and the message “Press button 1 to cancel.” The player will have 3 seconds to cancel it by pressing button 1 (previously used with another meaning). If canceled, the message “Movement canceled” will be written on the screen, and the previous board will be restored.
If the move is made, the board will be sent again via the serial line, and the turn will change to the next player.
The reception of the end-of-game command or pressing button 2 will indicate that the player whose turn it is concedes, and we will end the current game.
At the end of the game, the following information will be displayed:
The reason for ending (reset button, victory of one of the players, etc.)
Total processor usage time in this game (without power-down).
Total and average computation time of connect_K_hay_linea.
Total and average time it takes for a human to think about a move.
Total events queued in this game and a histogram by event type.
Additionally, the legend to play again will be shown. If the user decides to play again, a new game will start (with the initial empty board and alternating the turn of the starting player).
If the scheduler goes more than one second without processing any message or event, the system will reset. Considering that we are generating interrupts and periodic messages, this will only happen if the system is hung.
We want to continue maintaining the indications of overflow...